app/backend/requirements.txt
azure-identity==1.16.1
azure-search-documents==11.6.0b1
azure-storage-blob==12.23.1
Flask==3.0.3
Flask-Limiter==3.6.0
openai==1.27.0
numpy==1.26.4
isodate==0.6.1
chromadb==0.5.0
tiktoken==0.7.0
python-dotenv==1.1.1
httpx==0.27.2
requests==2.31.0
urllib3==2.1.0
